+++ HISTORY.rst
@@ -68,6 +68,7 @@
     - /api/v4/documents/{ ID }/versions/{ ID }/modify/
     - /api/v4/document_version_modification_backends/
 - Add workflow action to send user messages.
+- Update ``WorkflowAction`` to user ``common.classes.BaseBackend``.
 
 4.2.1 (2022-02-16)
 ==================